RoomView + Subject
Teile aus RoomView wurden zurück in Subject übertragen.
Außerdem werden die globalen TreeMaps verwendet.

Kleinere Anpassungen in RoomActivity(Wegen Übertragung der Animationen ins Subjekt)
und Room(Wenn für die Liste der im Raum vorhandenen Objekte null übergeben wird, wird eine leere Liste erzeugt).

Funktionalität scheint komplett vorhanden zu sein.
